requirements.txt
langchain
langchain-community
langchain-core
langchain-google-genai
openai
pypdf
tiktoken
aiofiles
fastapi
uvicorn
jinja2
python-multipart
faiss-cpu
python-dotenv
-e .